+++ pkger-core/src/lib.rs
@@ -7,13 +7,13 @@ extern crate lazy_static;
 pub mod archive;
 pub mod build;
 pub mod container;
-pub mod docker;
 pub mod gpg;
 pub mod image;
 #[macro_export]
 pub mod log;
 pub mod oneshot;
 pub mod recipe;
+pub mod runtime;
 pub mod ssh;
 pub mod template;